chore(STK-90): rename context folder into contexts for meaningfulness
