+++ packages/app/components/StacksTable.tsx
@@ -27,7 +27,7 @@ import {
   totalStacked,
 } from "@/models/stack-order";
 import { formatTokenValue } from "@/utils/token";
-import { ModalId, useModalContext } from "@/context";
+import { ModalId, useModalContext } from "@/contexts";
 
 export const StacksTable = ({ stackOrders }: { stackOrders: StackOrder[] }) => {
   const [stackOrder, setStackOrder] = useState<StackOrder>();
